requirements.txt
spotipy==2.23.0
streamlit==1.32.2
pandas==2.0.3
python-dotenv==1.0.0